rebase

+++ be/src/vec/aggregate_functions/aggregate_function_llm_agg.h
@@ -166,8 +166,7 @@ private:
 #endif
 
         return HttpClient::execute_with_retry(
-                static_cast<int>(_llm_config.max_retries),
-                static_cast<int>(_llm_config.retry_delay_second),
+                _llm_config.max_retries, _llm_config.retry_delay_second,
                 [this, &request_body, &response](HttpClient* client) -> Status {
                     return this->do_send_request(client, request_body, response);
                 });
